Fixed state machines and exiting the map editor, for bonuses
